--- a/requirements.txt
+++ b/requirements.txt
@@ -9,3 +9,4 @@ pytorch_lightning
 aiohttp
 accelerate
 pyyaml
+diffusers