--- a/requirements.txt
+++ b/requirements.txt
@@ -9,4 +9,3 @@ pytorch_lightning
 aiohttp
 accelerate
 pyyaml
-diffusers