* AITemplate Loader
  * Applies model to aitemplate module 
* KSampler (AITemplate)
  * Wraps AITemplate module for use with kdiffusion

Requires unet module compiled with [`compile_alt.py`](https://github.com/facebookincubator/AITemplate/blob/main/examples/05_stable_diffusion/scripts/compile_alt.py) and/or [`compile_controlnet.py`](https://github.com/facebookincubator/AITemplate/blob/main/examples/05_stable_diffusion/scripts/compile_controlnet.py)

[Collection of precompiled modules](https://huggingface.co/hlky/aitemplate)

Usage UNet:
`Load Checkpoint` -> `Load AITemplate` -> `KSampler (AITemplate)`

Usage ControlNet:
`Load ControlNet Model` -> `Load AITemplate (ControlNet)` -> `Apply ControlNet`
+load appropriate Control UNet


https://github.com/comfyanonymous/ComfyUI/assets/106811348/d76fd188-8d06-41ba-a84b-e1e8b8a0f64f

![image](https://github.com/comfyanonymous/ComfyUI/assets/106811348/31b0c9f5-c29b-43e8-a625-a3887e427b50)

Notes:
Needs ~~adjustment for 2.x support~~, ~~controlnet~~ etc may not work