@hlky 2023-05-15
Tested on a 4090 from runpod:tm:

~55it/s

50 steps

https://github.com/comfyanonymous/ComfyUI/assets/106811348/585931fc-a842-40d3-929d-a295d4477eea


20 steps


https://github.com/comfyanonymous/ComfyUI/assets/106811348/b203f8d5-b1f7-458b-889a-881cde6ecb35