@asagi4 2023-05-16
I gave this a quick test with your pre-compiled module and it is fast, but seems to ignore negative conditioning entirely, so output quality is terrible for me.

I'm not sure why the negative prompt gets ignored. The output only changes if I change the positive conditioning, and while negative conditioning does trigger a re-render, the output stays identical.