@asagi4 2023-05-16
That fixed it, indeed. The speedup is pretty crazy, especially for lower sizes. I generating 512x512 went from ~8.7it/s to 14.8 it/s. 768x768 went from 3.2it/s to 5it/s. (I have a 3060 too, on RHEL9)

Some other things I'm noticing with a bit more testing:
- Token merging doesn't seem to grant a speedup if AITemplate is used
- Batches larger than 1 throw an exception
- ControlNets don't cause errors, but they do nothing