@hlky 2023-05-16
>Token merging doesn't seem to grant a speedup if AITemplate is used

Tomesd is implemented in BasicTransformerBlock, I'll try adding it to AITemplate's BasicTransformerBlock but I'm not sure this will be able to work due to the compilation etc
>ControlNets don't cause errors, but they do nothing

ControlNet will require a module compiled for it and additional work on the AITemplateModelWrapper sampling side of things
>Batches larger than 1 throw an exception

The module I uploaded is only compiled for batch size 1, a module compiled for 64-1536 and up to batch size 4 is [here](https://huggingface.co/datasets/hlky/aitemplate/resolve/main/unet_bs.so)
As a general note; modules compiled for larger resolutions and larger batch size use more vram, but vram is consistent using the same amount regardless of the resolution and batch size of the generation