@ltdrdata 2023-05-17
I've tested on google colab. And I got this error message.

```
[01:53:43] model.h:29: Got error: no kernel image is available for execution on the device enum: 209 at constant_folder-generated.h: 109
[01:53:43] model_interface.cu:384: Error: Got error: no kernel image is available for execution on the device enum: 209 at constant_folder-generated.h: 109
```

![issue](https://github.com/comfyanonymous/ComfyUI/assets/128333288/ee3c9846-de45-4298-b82f-c4d216b5febe)