@hlky 2023-05-17
When the module is compiled, profilers are ran to determine the best algorithms to use. The modules I have shared are compiled on 3060/SM80+, so they won't work on SM75. Modules compiled on a SM75 gpu would work.