@hlky 2023-05-21
[ControlNet soon:tm:](https://github.com/facebookincubator/AITemplate/pull/713)