@hlky 2023-06-01
ControlNet support added, static sizes only for now

**ControlNet module must be used with specific ControlNet UNet**
**Using regular UNet module will not apply the control**

Modules for 512x512
[ControlNet](https://huggingface.co/datasets/hlky/aitemplate/resolve/main/controlnet_512_512.so)
[UNet](https://huggingface.co/datasets/hlky/aitemplate/resolve/main/unet_control_512_512.so)

Usage:
`Load ControlNet Model` -> `Load AITemplate (ControlNet)` -> `Apply ControlNet`