@jug-dev 2023-06-03
@comfyanonymous 

The individual behind the alias `hlky`, submitting this PR, has demonstrated actions that are against the spirit of open source, and against the sharing of stable diffusion based technology, I'm saddened that people like this exist in the world.

This individual issued a DCMA takedown notice against the project I started [hordelib](https://github.com/Haidra-Org/hordelib) which builds on top of ComfyUI, due to a cockup where one of the contributors removed a 4 letter alias from an AGPL copyright header. This DMCA takedown was issued without a single communication to me. Rather than simply reach out to me and discuss the issue which could have been easily resolved, they would rather attempt to takedown an entire volunteer online community which served to provide Stable Diffusion inference services to those who otherwise could not afford the expense of a modern GPU (i.e. Stable Horde).

https://github.com/github/dmca/blob/master/2023/05/2023-05-26-nataili.md

Prior to this, this individual had also deleted from Github another dependency that Stable Horde relied on in an attempt to take this community offline. 

This individual has demonstrated multiple actions seeking to shutdown stable diffusion based inference technology and specifically Stable Horde. It is obviously up to you if you think accepting submissions from such as individual is wise.