@hlky 2023-06-03
@jug-dev 

This pull request is unrelated to your issue, you are also free to reach out to me directly, alas you insist:

* db0 was instructed to cease using my code after I discovered a large portion of stable horde's traffic was CSAM
* eventually, hordelib is created, but my code is still used for several major components of hordelib
* however, license header is completely removed, this is in violation of the license agreement
* given the opportunity to resolve the matter, hordelib maintainers decide to mess around still and not reintroduce the original license header
* GitHub decided to takedown hordelib because maintainers did not comply

I have no intention of shutting down "stable diffusion based inference technology", I mean, you are literally commenting on a pull request where I've implemented AITemplate **for the community**.