@jshph 2023-06-13
@hlky which ControlNet model did you compile with AIT? I tried loading a binary which I compiled from `"lllyasviel/sd-controlnet-canny"`. However, this produces some errors. I get different errors when I load `lllyasviel/ControlNet-v1-1`.

I kind of wonder if this is due to how ComfyUI modifies the original lllyasviel ControlNet code whereas AIT compiles with diffusers/controlnet code.

Thanks a lot!