@hlky 2023-06-13
@jshph 

I compiled ControlNet with `[lllyasviel/sd-controlnet-canny](https://huggingface.co/lllyasviel/sd-controlnet-canny/)` but any model that works with Comfy should work, for example I've used [this one](https://huggingface.co/lllyasviel/ControlNet-v1-1/resolve/main/control_v11p_sd15_canny.pth).

Precompiled modules can be found on [huggingface](https://huggingface.co/datasets/hlky/aitemplate)

I've just pushed some fixes for ControlNet [here](https://github.com/hlky/AIT), there had been some changes to Comfy's controlnet code, that repo is the new home of my AITemplate code. If you still have issues feel free to create an issue over on that repo with the errors that you're experiencing.

Refer to the following image for how the graph should be set up. ControlNet is loaded as normal and fed into `Load AITemplate (ControlNet), a matching unet_control is also required, this needs the model fed from load checkpoint into `Load AITemplate` which then goes to KSampler.

![image](https://github.com/comfyanonymous/ComfyUI/assets/106811348/752a15eb-4cdc-4b5d-a347-723123efd49d)