@jshph 2023-06-14
Many thanks @hlky! I'll check out your work here, I'm sure it will be very helpful to get ControlNet/Comfy up and running.