@CyberTimon 2023-07-27
Can this be reimplemented for SD XL? Because SDXL is quite slow..