@2blackbar 2023-08-18
IS it me or it does not work with batches , works with only one image per generation